Refactor layout and default to showing replay info unless the a replay session exists
