+++ src/devtools/client/debugger/src/components/App.js
@@ -23,7 +23,8 @@ import {
   getSelectedSource,
 } from "../selectors";
 
-import { getSelectedPanel, getShowEditor } from "ui/reducers/app.ts";
+import { getSelectedPanel } from "ui/reducers/app";
+import { getShowEditor } from "ui/reducers/layout";
 import { useGetUserSettings } from "ui/hooks/settings";
 
 import KeyShortcuts from "devtools/client/shared/key-shortcuts";
