+++ src/ui/actions/comments.ts
@@ -4,12 +4,12 @@ import { actions } from "ui/actions";
 import { PendingComment, Comment, Reply, SourceLocation } from "ui/state/comments";
 import { UIThunkAction } from ".";
 import { ThreadFront } from "protocol/thread";
-import { setSelectedPrimaryPanel } from "./app";
 import escapeHtml from "escape-html";
 import { waitForTime } from "protocol/utils";
 import { PENDING_COMMENT_ID } from "ui/reducers/comments";
 import { RecordingId } from "@recordreplay/protocol";
 import { User } from "ui/types";
+import { setSelectedPrimaryPanel } from "./layout";
 const { getFilenameFromURL } = require("devtools/client/debugger/src/utils/sources-tree/getURL");
 const { getTextAtLocation } = require("devtools/client/debugger/src/reducers/sources");
 const { findClosestFunction } = require("devtools/client/debugger/src/utils/ast");
