+++ src/ui/setup/dynamic/devtools.ts
@@ -106,7 +106,6 @@ export default async (store: Store) => {
     app,
     comments,
     contextMenus,
-    layout,
     network,
     reactDevTools,
     timeline,
